+++ config/config_exec/protractor_conf/simple.conf.js
@@ -1,0 +1,4 @@
+var globalConf = require("../common.conf.js");
+globalConf.specs.push("../../../spec/TodoSpec.js");    
+globalConf.multiCapabilities= [{"browserName":("chrome")}],
+exports.config = globalConf;
