+++ helper/CalculationHelper.js
@@ -5,17 +5,8 @@ var myProfile = new FirefoxProfile();
 
 module.exports = function() {
 
-   this.get = function() {
-    browser.get("http://M1036232:Nov@2018@juliemr.github.io/protractor-demo/");
-   //  var capabilities = webdriver.Capabilities.firefox();
-   //  myProfile.setPreference("browser.newtab.url", "http://juliemr.github.io/protractor-demo/");
-   //  myProfile.encoded(function(err, encodedProfile) {
-   //          capabilities.set("firefox_profile", encodedProfile);
-   //  var wd = new webdriver.Builder().
-   //                    withCapabilities(capabilities).
-   //                    build();
-            
-   //  wd.get("http://juliemr.github.io/protractor-demo/");  });
+   this.get = async function() {
+      await browser.get("http://juliemr.github.io/protractor-demo/");
     };
 
    this.clickButton = function(locator) {
